third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/module/charset-01.html

<!DOCTYPE html>
<head>
  <meta charset="utf-8">
  <title>Module Scripts should always use UTF-8</title>
  <link rel="author" title="askalski" href="github.com/askalski">
  <link rel="help" href="https://html.spec.whatwg.org/multipage/scripting.html#attr-script-charset">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-utf8.js&ct=text/javascript&dummy=1">
  </script>
  <script type="module">
  test(function() {
    assert_equals(window.getSomeString(), "śćążź",
                  'Should be decoded as UTF-8');
  }, 'UTF-8 module script');
  </script>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-utf8.js&ct=text/javascript&dummy=2" charset="windows-1250">
  </script>
  <script type="module">
  test(function() {
    assert_equals(window.getSomeString(), "śćążź",
                  'Should be decoded as UTF-8');
  }, 'UTF-8 module script with wrong charset in attribute');
  </script>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-utf8.js&ct=text/javascript%3Bcharset=windows-1250&dummy=3">
  </script>
  <script type="module">
  test(function() {
    assert_equals(window.getSomeString(), "śćążź",
                  'Should be decoded as UTF-8');
  }, 'UTF-8 module script with wrong charset in Content-Type');
  </script>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-windows1250.js&ct=text/javascript&dummy=1">
  </script>
  <script type="module">
  test(function() {
    assert_not_equals(window.getSomeString(), "śćążź",
                      'Should be decoded as UTF-8');
  }, 'Non-UTF-8 module script');
  </script>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-windows1250.js&ct=text/javascript&dummy=2" charset="windows-1250">
  </script>
  <script type="module">
  test(function() {
    assert_not_equals(window.getSomeString(), "śćążź",
                      'Should be decoded as UTF-8');
  }, 'Non-UTF-8 module script with charset in attribute');
  </script>

  <script type="module" src="../serve-with-content-type.py?fn=external-script-windows1250.js&ct=text/javascript%3Bcharset=windows-1250">
  </script>
  <script type="module">
  test(function() {
    assert_not_equals(window.getSomeString(), "śćążź",
                      'Should be decoded as UTF-8');
  }, 'Non-UTF-8 module script with charset in Content-Type');
  </script>
</head>
